Move `Rate Us`, `Tell Friends` and `Contact Us` to MainSettings module
