+++ UnstoppableWallet/UnstoppableWallet/Modules/Settings/About/AboutModule.swift
@@ -1,12 +1,10 @@
 import UIKit
 
 struct AboutModule {
-
     static func viewController() -> UIViewController {
         let service = AboutService(
-                termsManager: App.shared.termsManager,
-                systemInfoManager: App.shared.systemInfoManager,
-                rateAppManager: App.shared.rateAppManager
+            termsManager: App.shared.termsManager,
+            systemInfoManager: App.shared.systemInfoManager
         )
         let releaseNotesService = ReleaseNotesService(appVersionManager: App.shared.appVersionManager)
 
@@ -14,5 +12,4 @@ struct AboutModule {
 
         return AboutViewController(viewModel: viewModel, urlManager: UrlManager(inApp: true))
     }
-
 }
